organize jsx files into folders that are reflected in the components.js file because it was just too much to scan through
