+++ app/assets/javascripts/components/collection/CollectionLink.jsx
