+++ app/assets/javascripts/components/collection/CollectionsList.jsx
