+++ app/assets/javascripts/components/collection/CollectionsListItem.jsx
