+++ app/assets/javascripts/components/item/AdditionalResources.jsx
