+++ app/assets/javascripts/components/item/AdditionalResourcesItem.jsx
