+++ app/assets/javascripts/components/item/ItemDescriptionTeaser.jsx
