+++ app/assets/javascripts/components/item/ItemsList.jsx
