+++ app/assets/javascripts/components/item/ItemsListItem.jsx
