+++ app/assets/javascripts/components/other/MetadataItem.jsx
