+++ app/assets/javascripts/components/other/MetadataList.jsx
