+++ app/assets/javascripts/components/other/OpenseadragonViewer.jsx
