+++ app/assets/javascripts/components/other/Thumbnail.jsx
