+++ app/assets/javascripts/components/section/NextSection.jsx
