+++ app/assets/javascripts/components/section/PreviousSection.jsx
