+++ app/assets/javascripts/components/section/Section.jsx
