+++ app/assets/javascripts/components/section/SectionCaption.jsx
