+++ app/assets/javascripts/components/section/SectionDescription.jsx
