+++ app/assets/javascripts/components/section/SectionImage.jsx
