+++ app/assets/javascripts/components/section/SectionLink.jsx
