+++ app/assets/javascripts/components/section/SectionShow.jsx
