+++ app/assets/javascripts/components/section/SectionsList.jsx
