+++ app/assets/javascripts/components/section/SectionsListItem.jsx
