+++ app/assets/javascripts/components/showcase/ShowcaseTitleSection.jsx
