+++ app/assets/javascripts/components/showcase/ShowcasesListItem.jsx
